Add http-server and manual test option
Adds the ability to 'npm run server' to allow browsing to
http://localhost:8080/ in order to do local testing and/or
debugging of fusebar.

In the process we add a top-level 'index.html' page and
replace the previous test JSON file with one that works
with the fields we use.

We've also added the 'http-server' as a development
dependency to enable this.

Signed-off-by: Daniel F. Dickinson <cshored@thecshore.com>