Separate taxonomy search from main
We now do a separate search on taxonomy terms (tags / categories) instead
of a single search.  Taxonomy terms are merged and pointed to the appropriate taxonomy page.
Content and title search remains largely the same, but we increase distance to improve search in long
text and with long patterns.  We also have a lower threshold to better limited the search results.

Signed-off-by: Daniel F. Dickinson <cshored@thecshore.com>